dataExtractor/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.2
kombu==5.3.5
langdetect==1.0.9
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
PyMuPDF==1.23.25
PyMuPDFb==1.23.22
pytesseract==0.3.10
python-dateutil==2.8.2
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
